Fixed ordering implementation for absolute and relative intervals
Basically the now-old Ord implementation was sorting by start and by decreasing end,
but this implementation produced the side-effect of being unreliable for checking for start equality,
which is an essential part for BTrees.

Therefore, the Ord implementation was rewritten to only consider the start bound of the intervals.
Moreover, Ord implementations were added for *Interval and Emptiable*Bounds (empty intervals are always lesser than the rest, but two empty intervals are equal)

The ordering by start and inverse end is implemented as the method ord_by_start_and_inv_length() as it can be useful for ordering intervals
when aware of the potential side-effects.