Replaced united sets with set operations on layered bounds
Since parallel iterators are extremely fast, i decided to remove united sets entirely,
as it used BTree for sorting and was only good for finding overlapping intervals amongst the set given a compared interval.
Otherwise, the system that was supposed to be running the set operations on united sets (which was the whole point of
having united sets in the first place) became convoluted, and some set operations even seemed almost impossible to
achieve in a reasonable manner (i.e. with reasonable efficiency)

Moreover, i definitely think that even if they could be done, they would have been hard to be made parallel.

In this commit the first set operation on layered bounds: intersection, is defined.