Removed precision iterator and relativity conversion iterator as they can be done through maps
Removed those iterators as they can easily be replaced with native Iterator methods that will provide
more control over the actual transformation. I realized that those iterators, although initially thought
as a convenience, were more of a weight as it would create two ways of doing the same thing, make so that
if the underlying trait (like ToAbsolute for relativity conversion iters) is changed, the iterator has to be changed,
etc.
After this reflection, I decided to remove those iterators.